+++ docs/src/references/changelog.rst
@@ -27,6 +27,7 @@ changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 Added
 #####
 
+* Updated to ``metatensor-torch`` version 0.7
 * Added classes for the calculation of dipole interactions
 * Better documentation for for ``cell``, ``charges`` and ``positions`` parameters